+++ rxjs-run/package.json
@@ -44,7 +44,7 @@
     "sinon": "^6.1.4",
     "ts-node": "^7.0.0",
     "tslint": "^5.11.0",
-    "typescript": "~2.8.4",
+    "typescript": "3.0.x",
     "xstream": "11.x"
   },
   "engines": {
