+++ rxjs-run/test/setup.ts
@@ -1,6 +1,7 @@
+// tslint:disable-next-line
+import 'mocha';
 import * as assert from 'assert';
-import * as sinon from 'sinon';
-import {run, setup} from '../src/index';
+import {setup} from '../src/index';
 import xs, {Stream} from 'xstream';
 import {Observable, of, from, range} from 'rxjs';
 import {take, startWith, map, delay, concatMap, tap} from 'rxjs/operators';
@@ -112,14 +113,14 @@ describe('setup', function() {
         delay(1)
       );
     }
-    const {sources, run: _run} = setup(app, {other: driver});
+    const {sources, run} = setup(app, {other: driver});
     let dispose: any;
     sources.other.subscribe(x => {
       assert.strictEqual(x, 97);
       dispose();
       done();
     });
-    dispose = _run();
+    dispose = run();
   });
 
   it('should not work after has been disposed', function(done) {
@@ -129,7 +130,7 @@ describe('setup', function() {
       return {other: number$};
     }
 
-    const {sources, run: _run} = setup(app, {
+    const {sources, run} = setup(app, {
       other: (num$: any) => from(num$).pipe(map((num: any) => 'x' + num)),
     });
 
@@ -143,100 +144,6 @@ describe('setup', function() {
         }, 100);
       }
     });
-    dispose = _run();
-  });
-});
-
-describe('run', function() {
-  it('should be a function', function() {
-    assert.strictEqual(typeof run, 'function');
-  });
-
-  it('should throw if first argument is not a function', function() {
-    assert.throws(() => {
-      (run as any)('not a function');
-    }, /First argument given to Cycle must be the 'main' function/i);
-  });
-
-  it('should throw if second argument is not an object', function() {
-    assert.throws(() => {
-      (run as any)(() => {}, 'not an object');
-    }, /Second argument given to Cycle must be an object with driver functions/i);
-  });
-
-  it('should throw if second argument is an empty object', function() {
-    assert.throws(() => {
-      (run as any)(() => {}, {});
-    }, /Second argument given to Cycle must be an object with at least one/i);
-  });
-
-  it('should return a dispose function', function() {
-    const sandbox = sinon.createSandbox();
-    const spy = sandbox.spy();
-    function app(sources: any): any {
-      return {
-        other: sources.other.pipe(
-          take(1),
-          startWith('a')
-        ),
-      };
-    }
-    function driver() {
-      return of('b').pipe(tap(spy));
-    }
-    const dispose = run(app, {other: driver});
-    assert.strictEqual(typeof dispose, 'function');
-    sinon.assert.calledOnce(spy);
-    dispose();
-  });
-
-  it('should report main() errors in the console', function(done) {
-    const sandbox = sinon.createSandbox();
-    sandbox.stub(console, 'error');
-
-    function main(sources: any): any {
-      const sink = sources.other.pipe(
-        take(1),
-        startWith('a'),
-        delay(10),
-        map(() => {
-          throw new Error('malfunction');
-        })
-      );
-
-      return {
-        other: sink,
-      };
-    }
-
-    function driver(xsSink: any) {
-      from(xsSink).subscribe({
-        next: () => {},
-        error: err => {},
-      });
-      return of('b');
-    }
-
-    let caught = false;
-    try {
-      run(main, {other: driver});
-    } catch (e) {
-      assert.strictEqual(e.message, 'malfunction');
-      caught = true;
-    }
-    setTimeout(() => {
-      sinon.assert.calledOnce(console.error as any);
-      sinon.assert.calledWithExactly(
-        console.error as any,
-        sinon.match((err: any) => err.message === 'malfunction')
-      );
-
-      // Should be false because the error was already reported in the console.
-      // Otherwise we would have double reporting of the error.
-      assert.strictEqual(caught, false);
-
-      sandbox.restore();
-      done();
-    }, 80);
+    dispose = run();
   });
 });
